client_bot/requirements.txt
python-telegram-bot==13.15
psycopg2-binary
python-dotenv
openai
